move postgresql code
